// frontend/src/shared/lib/api/model/provenanceBody.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * prot-cellar
 * OpenAPI spec version: 0.1.0
 */
import type { ProvenanceSourceType } from './provenanceSourceType';
import type { ProvenanceCitationBody } from './provenanceCitationBody';
import type { ProvenanceBodyContributorResearcher } from './provenanceBodyContributorResearcher';
import type { ProvenanceBodyObservedOn } from './provenanceBodyObservedOn';
import type { ProvenanceBodyNote } from './provenanceBodyNote';

export interface ProvenanceBody {
  source_type: ProvenanceSourceType;
  citations?: ProvenanceCitationBody[];
  contributor_researcher?: ProvenanceBodyContributorResearcher;
  observed_on?: ProvenanceBodyObservedOn;
  note?: ProvenanceBodyNote;
}