// frontend/src/shared/lib/api/model/provenanceSourceType.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * prot-cellar
 * OpenAPI spec version: 0.1.0
 */

export type ProvenanceSourceType = typeof ProvenanceSourceType[keyof typeof ProvenanceSourceType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ProvenanceSourceType = {
  published: 'published',
  preprint: 'preprint',
  private_comm: 'private_comm',
  internal: 'internal',
  patent: 'patent',
} as const;